@@ -1,37 +1,48 @@
-import { ErrorHandler, Inject, Injectable } from '@angular/core';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { I18NextPipe } from 'angular-i18next';
-import { asyncScheduler, never, Observable, scheduled, throwError } from 'rxjs';
+import { ErrorHandler, inject, Injectable } from '@angular/core';
+import { Router } from '@angular/router';
+import { I18NextService } from 'angular-i18next';
+import { Observable, of, throwError } from 'rxjs';
 import { LoggingService } from './logging.service';
+import { ToastsService } from './toast.service';
 
 // Handled exception handler, displays a toast/snackbar
 @Injectable({ providedIn: 'root' })
 export class ErrorHandlerService {
-  constructor(
-    @Inject(I18NextPipe) private i18NextPipe: I18NextPipe,
-    @Inject(MatSnackBar) private snackBar: MatSnackBar,
-    @Inject(LoggingService) private loggingService: LoggingService
-  ) {}
+  private router = inject(Router);
+  private toastsService = inject(ToastsService);
+  private i18NextService = inject(I18NextService);
+  private loggingService = inject(LoggingService);
 
   handle(error: string | Error): Observable<never> {
     const errorMessage = error instanceof Error ? error.message : error;
 
-    const translation = this.i18NextPipe.transform(`errors.${errorMessage}`);
+    const translation = this.i18NextService.t(`errors.${errorMessage}`);
     if (translation && translation !== errorMessage) {
       if (error !== 'unknown_exception') {
-        this.loggingService.logger.error(translation);
+        this.loggingService.error(translation);
       }
 
-      this.snackBar
-        .open(translation, 'Dismiss', {
-          duration: 3000,
-          panelClass: 'error',
-          horizontalPosition: 'right',
-        })
-        .onAction()
-        .subscribe(() => this.snackBar.dismiss());
+      if (error === 'unauthorized_access') {
+        if (this.router.url.startsWith('/sso')) {
+          return of();
+        }
+
+        if (this.router.url !== '/login') {
+          this.router.navigate(['login']);
+        }
+      }
+
+      if (this.router.url !== '/setup' && error === 'setup_required') {
+        this.router.navigate(['/setup']);
+      }
+
+      this.toastsService.show({
+        message: translation,
+        type: 'error',
+      });
+      console.error(translation);
     } else {
-      this.loggingService.logger.error(error);
+      this.loggingService.error(error);
       this.handle('unknown_exception');
     }
 
@@ -42,9 +53,11 @@ export class ErrorHandlerService {
 // Global unhandled exception handler
 @Injectable({ providedIn: 'root' })
 export class OverseerErrorHandler implements ErrorHandler {
-  constructor(private loggingService: LoggingService) {}
+  private loggingService = inject(LoggingService);
+  private errorHandlerService = inject(ErrorHandlerService);
 
   handleError(error: any): void {
-    this.loggingService.logger.error(error);
+    this.errorHandlerService.handle(error);
+    this.loggingService.error(error);
   }
 }
