@@ -1,91 +1,24 @@
-import { ApplicationConfig, importProvidersFrom, Provider, provideZoneChangeDetection } from '@angular/core';
-import { provideRouter } from '@angular/router';
-
-import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { ApplicationConfig, ErrorHandler, importProvidersFrom, provideZoneChangeDetection } from '@angular/core';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
+import { provideRouter } from '@angular/router';
 import { I18NextModule } from 'angular-i18next';
-import { NgxIndexedDBModule } from 'ngx-indexed-db';
-import { LoggerModule, NgxLoggerLevel } from 'ngx-logger';
-import { LocalStorageService, WebStorageModule } from 'ngx-store';
-import { environment } from '../environments/environment.local';
-import { dbConfig } from './app.db.config';
+import { modules, providers } from './app.config.local';
 import { routes } from './app.routes';
 import { I18N_PROVIDERS } from './app.translations';
-import { DialogService } from './dialogs/dialog.service';
-import { AuthenticationService } from './services/authentication.service';
-import { ControlService } from './services/control.service';
-import { LocalAuthenticationService } from './services/local/authentication.service';
-import { LocalControlService } from './services/local/control.service';
-import { IndexedStorageService } from './services/local/indexed-storage.service';
-import { LocalLoggingService } from './services/local/logging.service';
-import { LocalMachinesService } from './services/local/machines.service';
-import { LocalMonitoringService } from './services/local/monitoring.service';
-import { MachineProviderService } from './services/local/providers/machine-provider.service';
-import { LocalSettingsService } from './services/local/settings.service';
-import { LocalUsersService } from './services/local/users.service';
-import { LoggingService } from './services/logging.service';
-import { MachinesService } from './services/machines.service';
-import { MonitoringService } from './services/monitoring.service';
-import { RemoteAuthenticationService } from './services/remote/authentication.service';
-import { RemoteControlService } from './services/remote/control.service';
-import { RemoteLoggingService } from './services/remote/logging.service';
-import { RemoteMachinesService } from './services/remote/machines.service';
-import { RemoteMonitoringService } from './services/remote/monitoring.service';
-import { RemoteSettingsService } from './services/remote/settings.service';
-import { RemoteUsersService } from './services/remote/users.service';
-import { SettingsService } from './services/settings.service';
+import { OverseerErrorHandler } from './services/error-handler.service';
+import { LocalStorageService } from './services/local-storage.service';
 import { ThemeService } from './services/theme.service';
-import { UsersService } from './services/users.service';
-import { AuthenticationGuard } from './shared/authentication-guard';
-import { OverseerHttpInterceptor } from './app-http-interceptor';
-
-const services: Provider[] =
-  environment.serviceType === 'remote'
-    ? [
-        { provide: HTTP_INTERCEPTORS, useClass: OverseerHttpInterceptor, multi: true },
-        { provide: AuthenticationService, useClass: RemoteAuthenticationService },
-        { provide: ControlService, useClass: RemoteControlService },
-        { provide: LoggingService, useClass: RemoteLoggingService },
-        { provide: MachinesService, useClass: RemoteMachinesService },
-        { provide: MonitoringService, useClass: RemoteMonitoringService },
-        { provide: SettingsService, useClass: RemoteSettingsService },
-        { provide: UsersService, useClass: RemoteUsersService },
-      ]
-    : [
-        IndexedStorageService,
-        MachineProviderService,
-        { provide: AuthenticationService, useClass: LocalAuthenticationService },
-        { provide: ControlService, useClass: LocalControlService },
-        { provide: LoggingService, useClass: LocalLoggingService },
-        { provide: MachinesService, useClass: LocalMachinesService },
-        { provide: MonitoringService, useClass: LocalMonitoringService },
-        { provide: SettingsService, useClass: LocalSettingsService },
-        { provide: UsersService, useClass: LocalUsersService },
-      ];
-
-const modules = environment.serviceType === 'remote' ? [] : [NgxIndexedDBModule.forRoot(dbConfig)];
 
 export const appConfig: ApplicationConfig = {
   providers: [
-    provideZoneChangeDetection({ eventCoalescing: true }),
+    ...providers,
+    I18N_PROVIDERS,
+    ThemeService,
+    LocalStorageService,
     provideRouter(routes),
     provideAnimationsAsync(),
-    provideHttpClient(withInterceptorsFromDi()),
-    importProvidersFrom(
-      ...modules,
-      I18NextModule.forRoot(),
-      ThemeService,
-      WebStorageModule.forRoot(),
-      AuthenticationGuard,
-      LoggerModule.forRoot({
-        level: NgxLoggerLevel.DEBUG,
-        serverLogLevel: NgxLoggerLevel.INFO,
-      })
-    ),
-    I18N_PROVIDERS,
-    AuthenticationGuard,
-    LocalStorageService,
-    DialogService,
-    ...services,
+    provideZoneChangeDetection({ eventCoalescing: true }),
+    { provide: ErrorHandler, useClass: OverseerErrorHandler },
+    importProvidersFrom(...modules, I18NextModule.forRoot()),
   ],
 };
