@@ -1,25 +1,72 @@
 import { Routes } from '@angular/router';
-import { LoginComponent } from './login/login.component';
-import { SsoComponent } from './login/sso.component';
-import { MonitoringComponent } from './monitoring/monitoring.component';
-import { AuthenticationGuard } from './shared/authentication-guard';
+import { authenticationGuard } from './guards/authentication.guard';
+import { authorizationGuard } from './guards/authorization.guard';
 
 export const routes: Routes = [
   {
     path: '',
-    component: MonitoringComponent,
-    canActivate: [AuthenticationGuard],
+    pathMatch: 'full',
+    canActivate: [authenticationGuard()],
+    loadComponent: () => import('./pages/monitoring/monitoring.component').then((m) => m.MonitoringComponent),
   },
   {
     path: 'login',
-    component: LoginComponent,
+    loadComponent: () => import('./pages/login/login.component').then((m) => m.LoginComponent),
   },
   {
     path: 'sso',
-    component: SsoComponent,
+    loadComponent: () => import('./pages/sso/sso.component').then((m) => m.SsoComponent),
   },
   {
-    path: 'configuration',
-    loadChildren: () => import('./configuration/configuration.module').then((module) => module.ConfigurationModule),
+    path: 'setup',
+    loadComponent: () => import('./pages/setup/setup.component').then((m) => m.SetupComponent),
+  },
+  {
+    path: 'settings',
+    canActivate: [authenticationGuard(), authorizationGuard()],
+    loadComponent: () => import('./pages/configuration/configuration.component').then((m) => m.ConfigurationComponent),
+    children: [
+      {
+        path: '',
+        pathMatch: 'full',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/settings/settings.component').then((m) => m.SettingsComponent),
+      },
+      {
+        path: 'machines',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/list-machines/list-machines.component').then((m) => m.ListMachinesComponent),
+      },
+      {
+        path: 'machines/new',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/add-machine/add-machine.component').then((m) => m.AddMachineComponent),
+      },
+      {
+        path: 'machines/:id/edit',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/edit-machine/edit-machine.component').then((m) => m.EditMachineComponent),
+      },
+      {
+        path: 'users',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/list-users/list-users.component').then((m) => m.ListUsersComponent),
+      },
+      {
+        path: 'users/new',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/add-user/add-user.component').then((m) => m.AddUserComponent),
+      },
+      {
+        path: 'users/:id/edit',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/edit-user/edit-user.component').then((m) => m.EditUserComponent),
+      },
+      {
+        path: 'about',
+        canActivate: [authenticationGuard(), authorizationGuard()],
+        loadComponent: () => import('./pages/about/about.component').then((m) => m.AboutComponent),
+      },
+    ],
   },
 ];
