@@ -1,7 +1,7 @@
-export interface ApplicationInfo {
+export type ApplicationInfo = {
   platform?: string;
   operatingSystem?: string;
   machineName?: string;
   version?: string;
   runtime?: string;
-}
+};
