@@ -1,8 +1,8 @@
-export interface CertificateDetails {
+export type CertificateDetails = {
   id?: number;
   issuedTo?: string;
   issuedBy?: string;
   issuedDate?: string;
   expiredDate?: string;
   thumbprint?: string;
-}
+};
