@@ -2,18 +2,15 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map, mergeMap } from 'rxjs/operators';
 import { MachineToolType } from '../../models/machine.model';
-import { RequireAdministrator } from '../../shared/require-admin.decorator';
 import { ControlService } from '../control.service';
 import { MachinesService } from '../machines.service';
 import { MachineProviderService } from './providers/machine-provider.service';
 import { MachineProvider } from './providers/machine.provider';
+import { RequireAdministrator } from '../../decorators/require-admin.decorator';
 
 @Injectable({ providedIn: 'root' })
 export class LocalControlService implements ControlService {
-  constructor(
-    private machinesService: MachinesService,
-    private machineProviderService: MachineProviderService
-  ) {}
+  constructor(private machinesService: MachinesService, private machineProviderService: MachineProviderService) {}
 
   private getProvider(machineId: number): Observable<MachineProvider> {
     return this.machinesService.getMachine(machineId).pipe(map((machine) => this.machineProviderService.getProvider(machine)));
@@ -48,7 +45,7 @@ export class LocalControlService implements ControlService {
   setTemperature(machineId: number, heaterIndex: number, temperature: number): Observable<void> {
     return this.machinesService.getMachine(machineId).pipe(
       mergeMap((machine) => {
-        if (machine.tools.find((t) => t.toolType === MachineToolType.Heater && t.index === heaterIndex && t.name === 'bed')) {
+        if (machine.tools.find((t) => t.toolType === 'Heater' && t.index === heaterIndex && t.name === 'bed')) {
           return this.getProvider(machineId).pipe(mergeMap((provider) => provider.setBedTemperature(temperature)));
         }
 
