@@ -1,29 +1,26 @@
-import { Injectable } from '@angular/core';
-import { INGXLoggerMetadata, INGXLoggerMonitor, NGXLogger } from 'ngx-logger';
-import { LocalStorage } from 'ngx-store';
+import { inject, Injectable } from '@angular/core';
+import { map, Observable } from 'rxjs';
+import { LogEntry } from '../../models/log-entry.model';
 import { LoggingService } from '../logging.service';
 import { IndexedStorageService } from './indexed-storage.service';
 
-class OverseerMonitor implements INGXLoggerMonitor {
-  constructor(private storage: IndexedStorageService) {}
-
-  onLog(logObject: INGXLoggerMetadata): void {
-    this.storage.logging.add(logObject);
-  }
-}
-
 @Injectable({ providedIn: 'root' })
-export class LocalLoggingService implements LoggingService {
-  @LocalStorage() activeUser: any;
+export class LocalLoggingService extends LoggingService {
+  private indexedStorage = inject(IndexedStorageService);
 
-  get logger() {
-    return this.ngxLogger;
+  override download(): Observable<string> {
+    return this.indexedStorage.logging.getAll().pipe(
+      map((logEntries) => {
+        return logEntries
+          .map((e) => {
+            return `${e.timestamp} - ${e.level}: ${e.message}`;
+          })
+          .join('\n');
+      })
+    );
   }
 
-  constructor(
-    private ngxLogger: NGXLogger,
-    private storage: IndexedStorageService,
-  ) {
-    this.ngxLogger.registerMonitor(new OverseerMonitor(this.storage));
+  protected override saveLogEntry(entry: LogEntry): void {
+    this.indexedStorage.logging.add(entry).subscribe();
   }
 }
