@@ -1,19 +1,20 @@
-import { Injectable } from '@angular/core';
-import { NGXLogger, NgxLoggerLevel } from 'ngx-logger';
+import { inject, Injectable } from '@angular/core';
 import { LoggingService } from '../logging.service';
+import { LogEntry } from '../../models/log-entry.model';
+import { HttpClient } from '@angular/common/http';
+import { endpointFactory } from './endpoint-factory';
+import { map, Observable } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
-export class RemoteLoggingService implements LoggingService {
-  get logger() {
-    return this.ngxLogger;
+export class RemoteLoggingService extends LoggingService {
+  private getEndpoint = endpointFactory('/api/logging');
+  private http = inject(HttpClient);
+
+  protected override saveLogEntry(entry: LogEntry): void {
+    this.http.post(this.getEndpoint(), entry).subscribe();
   }
 
-  constructor(private ngxLogger: NGXLogger) {
-    this.ngxLogger.updateConfig({
-      level: NgxLoggerLevel.DEBUG,
-      serverLogLevel: NgxLoggerLevel.INFO,
-      disableConsoleLogging: true,
-      serverLoggingUrl: '/api/logging',
-    });
+  override download(): Observable<string> {
+    return this.http.get<{ content: string }>(this.getEndpoint()).pipe(map((response) => response.content));
   }
 }
