@@ -1,68 +1,20 @@
-import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
-import { LocalStorageService } from 'ngx-store';
-import { HttpClient } from '@angular/common/http';
-
-export interface ThemeUpdate {
-  current: string;
-  previous?: string;
-}
-
-export interface Theme {
-  name: string;
-  primaryColor: string;
-}
+import { effect, Injectable, signal } from '@angular/core';
+import { LocalStorageService } from './local-storage.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class ThemeService {
-  themes?: Theme[];
-  theme$: BehaviorSubject<ThemeUpdate>;
+  theme = signal<string | undefined>(undefined);
+  scheme = signal<'auto' | 'dark' | 'light'>('auto');
 
-  get currentTheme(): string {
-    return this.localStorage.get('currentTheme');
-  }
+  constructor(private localStorage: LocalStorageService) {
+    this.theme.set(this.localStorage.get('theme') ?? '#5cb248');
+    this.scheme.set(this.localStorage.get('scheme') ?? 'auto');
 
-  set currentTheme(value: string) {
-    this.localStorage.set('currentTheme', value);
-  }
-
-  get primaryColor() {
-    if (!this.themes) {
-      return '#fff';
-    }
-
-    const theme = this.themes.find((t) => t.name === this.currentTheme);
-    if (theme) {
-      return theme.primaryColor;
-    }
-
-    return this.themes.find((t) => t.name === 'default-theme')!.primaryColor;
-  }
-
-  constructor(
-    private localStorage: LocalStorageService,
-    http: HttpClient,
-  ) {
-    http
-      .get<Theme[]>('/themes.json')
-      .subscribe((themes) => (this.themes = themes));
-
-    const theme = this.currentTheme || 'default-theme';
-    this.theme$ = new BehaviorSubject<ThemeUpdate>({
-      current: theme,
+    effect(() => {
+      this.localStorage.set('theme', this.theme());
+      this.localStorage.set('scheme', this.scheme());
     });
-
-    this.currentTheme = theme;
-  }
-
-  applyTheme(themeClassName: string) {
-    this.theme$.next({
-      current: themeClassName,
-      previous: this.currentTheme,
-    });
-
-    this.currentTheme = themeClassName;
   }
 }
